split tests into node and browser  🐿 v2.5.16
