+++ Makefile
@@ -4,8 +4,13 @@ node_modules/@financial-times/n-gage/index.mk:
 
 -include node_modules/@financial-times/n-gage/index.mk
 
-unit-test:
+unit-test-node:
+	mocha test/node --recursive
+
+unit-test-browser:
 	karma start --single-run
+
+unit-test: unit-test-node unit-test-browser
 
 test: verify unit-test
