+++ karma.conf.js
@@ -6,10 +6,10 @@ module.exports = function (karma) {
 		files: [
 			// CircleCI's Chrome is version 43 at time of writing (no Object.assign support; latest is 47)
 			'https://cdn.polyfill.io/v2/polyfill.min.js?callback=ftNextInit&features=default&ua=chrome/43',
-			'tests/**/*.js'
+			'test/browser/**/*.js'
 		],
 		preprocessors: {
-			'tests/**/*.js': ['browserify']
+			'test/browser/**/*.js': ['browserify']
 		},
 		browserify: {
 			transform: ['babelify', 'debowerify', 'textrequireify'],
