+++ package.json
@@ -14,9 +14,10 @@
     "babelify": "^6.1.3",
     "bower": "^1.7.9",
     "browserify": "^11.0.1",
-    "chai": "^1.9.2",
+    "chai": "^4.1.1",
     "debowerify": "^1.3.1",
     "eslint": "^2.10.2",
+    "fetch-mock": "^5.12.2",
     "husky": "^0.14.1",
     "isomorphic-fetch": "^2.0.0",
     "karma": "^0.12.24",
@@ -27,9 +28,9 @@
     "karma-mocha": "^0.1.10",
     "karma-phantomjs-launcher": "0.1.2",
     "lintspaces-cli": "^0.1.1",
-    "mocha": "^2.0.1",
+    "mocha": "^3.5.0",
     "npm-prepublish": "^1.2.1",
-    "sinon": "^1.12.2",
+    "sinon": "^3.2.1",
     "textrequireify": "^2.1.1"
   },
   "scripts": {
