+++ test/node/lib/is-legacy-url.spec.js
@@ -1,8 +1,5 @@
-'use strict';
-
-import chai from 'chai';
-const expect = chai.expect;
-const isLegacyUrl = require('../../src/lib/is-legacy-url');
+const { expect } = require('chai');
+const isLegacyUrl = require('../../../src/lib/is-legacy-url');
 
 describe('identifying legacy URLs', () => {
 	it('should identify between legacy urls and current urls', () => {
